requirements.txt
pytest # for testing purposes
numpy
matplotlib
pillow # for image processing